+++ app/src/main/java/com/mertyigit0/vocabcards/ui/category/SharedViewModel.kt
@@ -3,5 +3,5 @@ package com.mertyigit0.vocabcards.ui.category
 import androidx.lifecycle.ViewModel
 
 class SharedViewModel : ViewModel() {
-    var categoryId: Long? = null
+    var categoryId: Int? = null
 }
